refactor(ui-auth): rewrite components to compiled Vertz JSX patterns (#1247)
Replace manual __child()/computed() primitives with declarative JSX and
let the Vertz compiler handle reactive transforms. This makes ui-auth
components follow the same patterns as user-written components.

- access-gate.tsx: JSX conditionals with ctx.accessSet reactivity
- auth-gate.tsx: JSX conditionals with ctx.status reactivity
- protected-route.tsx: JSX rendering + entitlement-check.ts helper for
  can() array checks (compiler can't track signal props in callbacks)
- user-name.tsx: direct <span> return, compiler handles ctx.user
- user-avatar.tsx: renamed static-branch var to avoid compiler collision
