+++ packages/ui-auth/src/user-avatar.tsx
@@ -1,7 +1,6 @@
 import { useContext } from '@vertz/ui';
 import type { User } from '@vertz/ui/auth';
 import { AuthContext, getUserDisplayName } from '@vertz/ui/auth';
-import { __child } from '@vertz/ui/internals';
 import type { JSX } from '@vertz/ui/jsx-runtime';
 import { Avatar } from './avatar';
 
@@ -19,10 +18,10 @@ export function UserAvatar({
   class: className,
 }: UserAvatarProps): JSX.Element {
   if (user) {
-    const avatarUrl = typeof user.avatarUrl === 'string' ? user.avatarUrl : undefined;
+    const src = typeof user.avatarUrl === 'string' ? user.avatarUrl : undefined;
     return (
       <Avatar
-        src={avatarUrl}
+        src={src}
         alt={getUserDisplayName(user)}
         size={size}
         fallback={fallback}
@@ -36,10 +35,9 @@ export function UserAvatar({
     throw new Error('UserAvatar must be used within AuthProvider, or pass a `user` prop');
   }
 
-  return __child(() => {
-    const avatarUrl =
-      ctx.user && typeof ctx.user.avatarUrl === 'string' ? ctx.user.avatarUrl : undefined;
-    const alt = getUserDisplayName(ctx.user);
-    return <Avatar src={avatarUrl} alt={alt} size={size} fallback={fallback} class={className} />;
-  });
+  const avatarUrl =
+    ctx.user && typeof ctx.user.avatarUrl === 'string' ? ctx.user.avatarUrl : undefined;
+  const alt = getUserDisplayName(ctx.user);
+
+  return <Avatar src={avatarUrl} alt={alt} size={size} fallback={fallback} class={className} />;
 }
