+++ packages/ui-auth/src/user-name.tsx
@@ -1,7 +1,6 @@
 import { useContext } from '@vertz/ui';
 import type { User } from '@vertz/ui/auth';
 import { AuthContext, getUserDisplayName } from '@vertz/ui/auth';
-import { __child } from '@vertz/ui/internals';
 import type { JSX } from '@vertz/ui/jsx-runtime';
 
 export interface UserNameProps {
@@ -24,7 +23,5 @@ export function UserName({
     throw new Error('UserName must be used within AuthProvider, or pass a `user` prop');
   }
 
-  return __child(() => {
-    return <span class={className}>{getUserDisplayName(ctx.user, fallback)}</span>;
-  });
+  return <span class={className}>{getUserDisplayName(ctx.user, fallback)}</span>;
 }
